max_octaves needs to be int.
Clang will throw error when trying to compile,
the socket for this is an IntSocket anyway.

+++ src/render/rhino_nodes.h
@@ -215,7 +215,7 @@ class RhinoFbmTextureNode : public ShaderNode {
   float alpha2;
 
   bool is_turbulent;
-  float max_octaves;
+  int max_octaves;
   float gain;
   float roughness;
